\name{doubexp}
\alias{doubexp}
\title{
Double exponential Link Function
}
\description{
Computes the double exponential link function, including its inverse.
}
\usage{
doubexp()
}
\value{
The double exponential transformation of theta.
}
\references{
Ford I, Torsney B, Wu C (1992). "The Use of a Canonical Form in the Construction 
of Locally Optimal Designs for Non-linear Problems." Journal of the Royal 
Statistical Society B, 54, 569-583. \url{http://www.jstor.org/stable/2346142}
}
\author{
David M. Smith <smithdm1@us.ibm.com>
}
\keyword{ models }
