\name{doubrecip}
\alias{doubrecip}
\title{
Double reciprocal Link Function
}
\description{
Computes the double reciprocal link function, including its inverse.
}
\usage{
doubrecip()
}
\value{
The double reciprocal transformation of theta.
}
\references{
Ford I, Torsney B, Wu C (1992). "The Use of a Canonical Form in the Construction 
of Locally Optimal Designs for Non-linear Problems." Journal of the Royal 
Statistical Society B, 54, 569-583. \url{http://www.jstor.org/stable/2346142}
}
\author{
David M. Smith <david.m.smith@truvenhealth.com>
}
\keyword{ models }
