\name{fitted.BinaryEPPM}
\alias{fitted.BinaryEPPM}
\title{
Extraction of fitted values from BinaryEPPM Objects
}
\description{
This function is generic. Extract the fitted values from models of class "BinaryEPMM".
}
\usage{
\method{fitted}{BinaryEPPM}(object, \dots)
}
\arguments{
  \item{object}{
fitted model object of class "BinaryEPPM".
}
  \item{\dots}{
currently not used.
}
}
\seealso{
\code{\link[stats:stats-package]{fitted}} 
}
\author{
David M. Smith <david.m.smith@truvenhealth.com>
}
\keyword{ methods }