\name{loglog}
\alias{loglog}
\title{
Log-log Link Function
}
\description{
Computes the loglog link function, including its inverse.
}
\usage{
loglog()
}
\value{
The loglog of theta where the logarithms are to base e.
}
\details{
Same link function as in Cribari-Neto and Zeileis (2010).
}
\references{
Cribari-Neto F, Zeileis A. (2010). Beta Regression in R.
\emph{Journal of Statistical Software}, \bold{34}(2), 1-24. 
\doi{10.18637/jss.v034.i02}.
}
\author{
David M. Smith <smithdm1@us.ibm.com>
}
\keyword{ models }
