\name{negcomplog}
\alias{negcomplog}
\title{
Negative complementary log-log Link Function
}
\description{
Computes the negative complementary log-log link function, including its inverse.
}
\usage{
negcomplog()
}
\value{
The negative complementary log-log of theta.
}
\references{
Tibshirani RJ, Ciampi A (1983). "A Family of Proportional- and Additive-Hazards Models for Survival Data". Biometrics 39(1), 141-147.
}
\author{
David M. Smith <david.m.smith@truvenhealth.com>
}
\keyword{ models }
