\name{Williams.litters}
\alias{Williams.litters}
\docType{data}
\title{
Number of implantations, data of Williams (1996).
}
\description{
The data is arranged as a list of frequency distributions of numbers of surviving foetuses for an given litter size. The litter size is the length of the frequency distribution. Dose is included both as a variate (vdose) and as a factor (fdose).
}
\usage{data("Williams.litters")}
\format{
  The format is:
List of 3
 $ vdose           : num [1:34] 0.75 1.5 1.5 0 0.75 1.5 3 0.75 1.5 3 ...
 $ fdose           : Factor w/ 4 levels "0","0.75","1.5",..: 2 3 3 1 2 3 4 2 3 4 ...
 $ number.surviving:List of 34
  ..$ : num [1:8] 0 0 0 0 0 0 0 1
  ..$ : num [1:8] 0 0 0 1 0 0 0 0
  ..$ : num [1:9] 0 0 0 0 0 0 0 0 1
  ..$ : num [1:10] 0 0 0 0 0 0 0 1 1 0
  ..$ : num [1:10] 0 0 0 0 0 0 0 1 0 0
  ..$ : num [1:10] 0 0 0 0 0 0 0 2 0 0
  ..$ : num [1:10] 0 0 0 0 1 0 0 2 0 0
  ..$ : num [1:11] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:11] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:11] 0 0 0 0 1 0 0 0 0 1 ...
  ..$ : num [1:12] 0 0 0 0 0 0 1 0 0 0 ...
  ..$ : num [1:12] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ : num [1:12] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ : num [1:13] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:13] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:13] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ : num [1:13] 0 0 0 0 0 0 0 1 0 0 ...
  ..$ : num [1:14] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:14] 0 0 0 0 0 0 0 0 1 0 ...
  ..$ : num [1:14] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:14] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:15] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:15] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:15] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:15] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:16] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:16] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:16] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:16] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:17] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:17] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:17] 0 0 0 0 0 0 0 1 0 0 ...
  ..$ : num [1:18] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:20] 0 0 0 0 0 0 0 0 0 0 ...
}
\source{
Williams D (1996). "Overdispersion in Logistic Linear Models." In B Morgan (ed.), 
Statistics in Toxicology, pp75-84. Oxford Science Publications.
}
\references{
Faddy M, Smith D. (2012). Extended Poisson Process Modeling and 
Analysis of Grouped Binary Data. \emph{Biometrical Journal}, \bold{54}, 426-435.
\doi{10.1002/bimj.201100214}.
}
\examples{
data("Williams.litters")
}
\keyword{datasets}