\name{doubrecip}
\alias{doubrecip}
\title{
Double reciprocal Link Function
}
\description{
Computes the double reciprocal link function, including its inverse.
}
\usage{
doubrecip()
}
\value{
The double reciprocal transformation of theta.
}
\references{
Ford I, Torsney B, Wu C (1992). "The Use of a Canonical Form in the Construction 
of Locally Optimal Designs for Non-linear Problems." Journal of the Royal 
Statistical Society B, 54, 569-583. \doi{10.1111/j.2517-6161.1992.tb01897.x}
}
\author{
David M. Smith <smithdm1@us.ibm.com>
}
\keyword{ models }
